Fix: Allows for background scrolling, refactored the socket issue.
