+++ src/contexts/Socket/index.jsx
@@ -3,6 +3,6 @@ import { io } from 'socket.io-client';
 import { SERVICES_ENDPOINT } from 'constant';
 
 export const socket = io(SERVICES_ENDPOINT);
-const SocketContext = createContext({ socket });
+const SocketContext = createContext(socket);
 
 export default SocketContext;
